+++ kvlds/dispatch.c
@@ -614,7 +614,7 @@ dispatch_accept(int s, struct btree * T,
 	D->mr_timer = NULL;
 	D->mr_timer_expired = 0;
 	D->mr_timeout.tv_sec = (time_t)w;
-	D->mr_timeout.tv_usec = (suseconds_t)((w - D->mr_timeout.tv_sec)
+	D->mr_timeout.tv_usec = (suseconds_t)((w - (double)D->mr_timeout.tv_sec)
 	    * 1000000);
 	D->mr_min_batch = g;
